Rewrite controller
