--- controller/Judger.js
@@ -1,5 +1,0 @@
-import Judger from "../driver/kon";
-// TODO: Make Judgers singleton
-export const Judgers = ["http://localhost:30000"].map(
-    (host) => new Judger(host)
-);
